requirements.txt
Flask==2.1.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
Jinja2==3.1.1
python-dotenv==0.20.0
requests==2.11.1
SQLAlchemy==1.4.35
Werkzeug==2.1.0
WTForms==3.0.1
